Move some of WinRT collections tests to their respective packages
